add rating&baseEntity&history

+++ src/main/java/com/softwaredesign/recruitservice/repository/RatingRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RatingRepository extends GenericRepository<Rating, Integer> {
+public interface RatingRepository extends GenericRepository<Rating, Long> {
 
     Rating findRatingById(Long id);
 }
